deprecate processor methods in obsreport
These deprecated methods/structs have been moved to processorhelper:
  - `obsreport.BuildProcessorCustomMetricName` -> `processorhelper.BuildCustomMetricName`
  - `obsreport.Processor` -> `processorhelper.Processor`
  - `obsreport.ProcessorSettings` -> `processorhelper.Settings`
  - `obsreport.NewProcessor` -> `processorhelper.New`

Same as the change for the exporter, no functional changes just moving code over to the
new location.
