+++ obsreport/obsreport_test.go
@@ -25,9 +25,8 @@ const (
 )
 
 var (
-	receiverID  = component.NewID("fakeReceiver")
-	scraperID   = component.NewID("fakeScraper")
-	processorID = component.NewID("fakeProcessor")
+	receiverID = component.NewID("fakeReceiver")
+	scraperID  = component.NewID("fakeScraper")
 
 	errFake        = errors.New("errFake")
 	partialErrFake = scrapererror.NewPartialScrapeError(errFake, 1)
@@ -309,82 +308,4 @@ func TestReceiveWithLongLivedCtx(t *testing.T) {
 			t.Fatalf("unexpected error: %v", params[i].err)
 		}
 	}
-}
-
-func TestProcessorTraceData(t *testing.T) {
-	testTelemetry(t, processorID, func(t *testing.T, tt obsreporttest.TestTelemetry, useOtel bool) {
-		const acceptedSpans = 27
-		const refusedSpans = 19
-		const droppedSpans = 13
-		obsrep, err := newProcessor(ProcessorSettings{
-			ProcessorID:             processorID,
-			ProcessorCreateSettings: tt.ToProcessorCreateSettings(),
-		}, useOtel)
-		require.NoError(t, err)
-		obsrep.TracesAccepted(context.Background(), acceptedSpans)
-		obsrep.TracesRefused(context.Background(), refusedSpans)
-		obsrep.TracesDropped(context.Background(), droppedSpans)
-
-		require.NoError(t, tt.CheckProcessorTraces(acceptedSpans, refusedSpans, droppedSpans))
-	})
-}
-
-func TestProcessorMetricsData(t *testing.T) {
-	testTelemetry(t, processorID, func(t *testing.T, tt obsreporttest.TestTelemetry, useOtel bool) {
-		const acceptedPoints = 29
-		const refusedPoints = 11
-		const droppedPoints = 17
-
-		obsrep, err := newProcessor(ProcessorSettings{
-			ProcessorID:             processorID,
-			ProcessorCreateSettings: tt.ToProcessorCreateSettings(),
-		}, useOtel)
-		require.NoError(t, err)
-		obsrep.MetricsAccepted(context.Background(), acceptedPoints)
-		obsrep.MetricsRefused(context.Background(), refusedPoints)
-		obsrep.MetricsDropped(context.Background(), droppedPoints)
-
-		require.NoError(t, tt.CheckProcessorMetrics(acceptedPoints, refusedPoints, droppedPoints))
-	})
-}
-
-func TestBuildProcessorCustomMetricName(t *testing.T) {
-	tests := []struct {
-		name string
-		want string
-	}{
-		{
-			name: "firstMeasure",
-			want: "processor/test_type/firstMeasure",
-		},
-		{
-			name: "secondMeasure",
-			want: "processor/test_type/secondMeasure",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := BuildProcessorCustomMetricName("test_type", tt.name)
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func TestProcessorLogRecords(t *testing.T) {
-	testTelemetry(t, processorID, func(t *testing.T, tt obsreporttest.TestTelemetry, useOtel bool) {
-		const acceptedRecords = 29
-		const refusedRecords = 11
-		const droppedRecords = 17
-
-		obsrep, err := newProcessor(ProcessorSettings{
-			ProcessorID:             processorID,
-			ProcessorCreateSettings: tt.ToProcessorCreateSettings(),
-		}, useOtel)
-		require.NoError(t, err)
-		obsrep.LogsAccepted(context.Background(), acceptedRecords)
-		obsrep.LogsRefused(context.Background(), refusedRecords)
-		obsrep.LogsDropped(context.Background(), droppedRecords)
-
-		require.NoError(t, tt.CheckProcessorLogs(acceptedRecords, refusedRecords, droppedRecords))
-	})
 }
