+++ processor/batchprocessor/metrics.go
@@ -17,8 +17,8 @@ import (
 
 	"go.opentelemetry.io/collector/config/configtelemetry"
 	"go.opentelemetry.io/collector/internal/obsreportconfig/obsmetrics"
-	"go.opentelemetry.io/collector/obsreport"
 	"go.opentelemetry.io/collector/processor"
+	"go.opentelemetry.io/collector/processor/processorhelper"
 )
 
 const (
@@ -50,7 +50,7 @@ func metricViews() []*view.View {
 	processorTagKeys := []tag.Key{processorTagKey}
 
 	countBatchSizeTriggerSendView := &view.View{
-		Name:        obsreport.BuildProcessorCustomMetricName(typeStr, statBatchSizeTriggerSend.Name()),
+		Name:        processorhelper.BuildCustomMetricName(typeStr, statBatchSizeTriggerSend.Name()),
 		Measure:     statBatchSizeTriggerSend,
 		Description: statBatchSizeTriggerSend.Description(),
 		TagKeys:     processorTagKeys,
@@ -58,7 +58,7 @@ func metricViews() []*view.View {
 	}
 
 	countTimeoutTriggerSendView := &view.View{
-		Name:        obsreport.BuildProcessorCustomMetricName(typeStr, statTimeoutTriggerSend.Name()),
+		Name:        processorhelper.BuildCustomMetricName(typeStr, statTimeoutTriggerSend.Name()),
 		Measure:     statTimeoutTriggerSend,
 		Description: statTimeoutTriggerSend.Description(),
 		TagKeys:     processorTagKeys,
@@ -66,7 +66,7 @@ func metricViews() []*view.View {
 	}
 
 	distributionBatchSendSizeView := &view.View{
-		Name:        obsreport.BuildProcessorCustomMetricName(typeStr, statBatchSendSize.Name()),
+		Name:        processorhelper.BuildCustomMetricName(typeStr, statBatchSendSize.Name()),
 		Measure:     statBatchSendSize,
 		Description: statBatchSendSize.Description(),
 		TagKeys:     processorTagKeys,
@@ -74,7 +74,7 @@ func metricViews() []*view.View {
 	}
 
 	distributionBatchSendSizeBytesView := &view.View{
-		Name:        obsreport.BuildProcessorCustomMetricName(typeStr, statBatchSendSizeBytes.Name()),
+		Name:        processorhelper.BuildCustomMetricName(typeStr, statBatchSendSizeBytes.Name()),
 		Measure:     statBatchSendSizeBytes,
 		Description: statBatchSendSizeBytes.Description(),
 		TagKeys:     processorTagKeys,
@@ -140,35 +140,35 @@ func (bpt *batchProcessorTelemetry) createOtelMetrics(mp metric.MeterProvider, c
 	meter := mp.Meter(scopeName)
 
 	bpt.batchSizeTriggerSend, err = meter.Int64Counter(
-		obsreport.BuildProcessorCustomMetricName(typeStr, "batch_size_trigger_send"),
+		processorhelper.BuildCustomMetricName(typeStr, "batch_size_trigger_send"),
 		metric.WithDescription("Number of times the batch was sent due to a size trigger"),
 		metric.WithUnit("1"),
 	)
 	errors = multierr.Append(errors, err)
 
 	bpt.timeoutTriggerSend, err = meter.Int64Counter(
-		obsreport.BuildProcessorCustomMetricName(typeStr, "timeout_trigger_send"),
+		processorhelper.BuildCustomMetricName(typeStr, "timeout_trigger_send"),
 		metric.WithDescription("Number of times the batch was sent due to a timeout trigger"),
 		metric.WithUnit("1"),
 	)
 	errors = multierr.Append(errors, err)
 
 	bpt.batchSendSize, err = meter.Int64Histogram(
-		obsreport.BuildProcessorCustomMetricName(typeStr, "batch_send_size"),
+		processorhelper.BuildCustomMetricName(typeStr, "batch_send_size"),
 		metric.WithDescription("Number of units in the batch"),
 		metric.WithUnit("1"),
 	)
 	errors = multierr.Append(errors, err)
 
 	bpt.batchSendSizeBytes, err = meter.Int64Histogram(
-		obsreport.BuildProcessorCustomMetricName(typeStr, "batch_send_size_bytes"),
+		processorhelper.BuildCustomMetricName(typeStr, "batch_send_size_bytes"),
 		metric.WithDescription("Number of bytes in batch that was sent"),
 		metric.WithUnit("By"),
 	)
 	errors = multierr.Append(errors, err)
 
 	bpt.batchMetadataCardinality, err = meter.Int64ObservableUpDownCounter(
-		obsreport.BuildProcessorCustomMetricName(typeStr, "metadata_cardinality"),
+		processorhelper.BuildCustomMetricName(typeStr, "metadata_cardinality"),
 		metric.WithDescription("Number of distinct metadata value combinations being processed"),
 		metric.WithUnit("1"),
 		metric.WithInt64Callback(func(_ context.Context, obs metric.Int64Observer) error {
